+++ addons/l10n_br/__init__.py
@@ -1,3 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
+from . import demo
 from . import models
 from . import wizard
