+++ addons/l10n_br/demo/__init__.py
@@ -1,0 +1,1 @@
+from . import account_demo
