+++ addons/l10n_br/views/account_journal_views.xml
@@ -6,7 +6,7 @@
             <field name="name">account.journal.form</field>
             <field name="inherit_id" ref="l10n_latam_invoice_document.view_account_journal_form"/>
             <field name="arch" type="xml">
-                 <field name="type" position="after">
+                <field name="type" position="after">
                     <field name="l10n_br_invoice_serial" invisible="not l10n_latam_use_documents or country_code != 'BR'"/>
                 </field>
             </field>
